Загрузка тестовых данных в Elasticsearch

+++ tests/tests/functional/utils/elasticsearch/indices.py
@@ -1,5 +1,7 @@
 from __future__ import annotations
 
+from collections.abc import Iterable
+
 import elasticsearch
 import elasticsearch.helpers
 
@@ -9,7 +11,7 @@ class ElasticsearchIndex:
     index_name: str
     index_data: dict
 
-    index_created: bool
+    _index_created: bool
 
     def __init__(self,
                  *,
@@ -20,9 +22,12 @@ class ElasticsearchIndex:
         self.index_name = index_name
         self.index_data = index_data
 
-        self.index_created = False
+        self._index_created = False
 
-    async def create_index(self) -> None:
+    async def create_index(self, *, recreate=False) -> None:
+        if self._index_created and not recreate:
+            return
+
         await self.delete_index()
 
         try:
@@ -31,7 +36,7 @@ class ElasticsearchIndex:
             if e.error == 'resource_already_exists_exception':
                 pass
 
-        self.index_created = True
+        self._index_created = True
 
     async def delete_index(self) -> None:
         try:
@@ -40,4 +45,12 @@ class ElasticsearchIndex:
             if e.error == 'index_not_found_exception':
                 pass
 
-        self.index_created = False
+        self._index_created = False
+
+    async def load_data(self, *, documents: Iterable[dict]) -> None:
+        await self.create_index()
+        await elasticsearch.helpers.async_bulk(self.client, ({
+            '_index': self.index_name,
+            '_id': document_data['id'],
+            '_source': document_data,
+        } for document_data in documents), refresh='wait_for')
